requirements.txt
# Pinned versions — prevents silent breaking upgrades on a production kiosk.
# To upgrade: test in a staging venv first.
flask==3.0.3
flask-socketio==5.3.6
adafruit-circuitpython-pn532==2.1.3
gspread==6.1.2
google-auth==2.29.0
bcrypt==4.1.3
itsdangerous==2.2.0
